+++ packages/resourceful-redux/src/utils/cru-reducer-helper.js
@@ -1,22 +1,36 @@
 import setResourceMeta from './set-resource-meta';
 import upsertResources from './upsert-resources';
 import requestStatuses from './request-statuses';
+import warning from './warning';
 
 // This reducer helper handles the "CRU" in "CRUD".
 
 export default function(state, action, {initialResourceMeta}, updatedMeta) {
   const resources = action.resources;
-  const label = action.label;
   const resourcesIsUndefined = typeof resources === 'undefined';
   const hasResources = resources && resources.length;
 
+  let label;
+  if (action.label && typeof action.label === 'string') {
+    label = action.label;
+  }
+
   if (process.env.NODE_ENV !== 'production') {
     if (!resources) {
-      console.warn(
-        `A 'resources' Array was not supplied to a Resourceful Redux "success" ` +
-        `action with type "${action.type}". Without a 'resources' Array, ` +
-        `Resourceful Redux will not be able to track which resources have been ` +
-        `affected by this CRUD operation. You should check your Action Creators.`
+      warning(
+        `A 'resources' array was not included in a Resourceful Redux ` +
+        `"success" action with type "${action.type}. Without a 'resources' ` +
+        `Array, Resourceful Redux will not be able to track which resources ` +
+        `were affected by this CRUD operation. You should check your Action ` +
+        `Creators to make sure that they always include a 'resources' array.`
+      );
+    } else if (!Array.isArray(resources)) {
+      warning(
+        `A non-array 'resources' value was passed to a Resourceful Redux ` +
+        `"success" action with type "${action.type}". 'resources' must be an ` +
+        `array. If your backend returned a single object, be sure to wrap it ` +
+        `inside of an array. If you're using the Resourceful Action Creators ` +
+        `library, you can do this using the "transformData" option.`
       );
     }
   }