+++ packages/resourceful-redux/test/unit/reducers/create.js
@@ -1,3 +1,42 @@
+import {resourceReducer} from '../../../src';
+
 describe('reducers: create', function() {
-  // Intentionally blank. See reducers: read tests
+  // Intentionally mostly-empty. Refer to the read reducer tests and the
+  // reducer-generator tests
+
+  describe('CREATE_RESOURCES_SUCCEEDED:', () => {
+    it('returns the right state without a label, without IDs', () => {
+      stub(console, 'error');
+      const initialState = {
+        resources: {
+          1: {id: 1},
+          3: {id: 3},
+          4: {id: 4},
+        },
+        labels: {
+          pasta: {
+            hungry: true
+          }
+        },
+        meta: {
+          1: {
+            name: 'what'
+          },
+          3: {
+            deleteStatus: 'sandwiches'
+          }
+        }
+      };
+
+      const reducer = resourceReducer('hellos', {initialState});
+
+      const reduced = reducer(undefined, {
+        type: 'CREATE_RESOURCES_SUCCEEDED',
+        resourceName: 'hellos'
+      });
+
+      expect(reduced).to.deep.equal(initialState);
+      expect(console.error.callCount).to.equal(1);
+    });
+  });
 });