+++ packages/resourceful-redux/test/unit/reducers/index.js
@@ -7,6 +7,7 @@ describe('reducer', function() {
   });
 
   it('should handle an action type that does not exist', () => {
+    stub(console, 'error');
     const reducer = resourceReducer('hellos');
 
     const state = {
@@ -16,5 +17,18 @@ describe('reducer', function() {
 
     const reduced = reducer(state, {type: 'does_not_exist'});
     expect(reduced).to.equal(state);
+    expect(console.error.callCount).to.equal(0);
+  });
+
+  it('should warn when no resourceName is passed', () => {
+    stub(console, 'error');
+    const reducer = resourceReducer();
+    expect(console.error.callCount).to.equal(1);
+  });
+
+  it('should warn when a non-string resourceName is passed', () => {
+    stub(console, 'error');
+    const reducer = resourceReducer({});
+    expect(console.error.callCount).to.equal(1);
   });
 });