+++ packages/resourceful-redux/test/unit/reducers/plugins.js
@@ -1,10 +1,22 @@
 import {resourceReducer, actionTypes} from '../../../src';
 
 describe('reducer', function() {
+  it('should warn when a bad plugin is initialized', () => {
+    stub(console, 'error');
+
+    const reducer = resourceReducer('hellos', {
+      plugins: [
+        () => {}
+      ]
+    });
+
+    expect(console.error.callCount).to.equal(1);
+  });
+
   it('should handle a plug-in on a built-in type', () => {
     const reducer = resourceReducer('hellos', {
       plugins: [
-        (state, action) => {
+        () => (state, action) => {
           if (action.type === actionTypes.READ_RESOURCES_SUCCEEDED) {
             return {
               ...state,
@@ -43,7 +55,7 @@ describe('reducer', function() {
   it('should handle a plug-in on a custom type', () => {
     const reducer = resourceReducer('hellos', {
       plugins: [
-        (state, action) => {
+        () => (state, action) => {
           if (action.type === 'SANDWICHES_ARE_GOOD') {
             return {
               ...state,
@@ -73,7 +85,7 @@ describe('reducer', function() {
   it('should handle multiple plug-ins on a custom type, from right to left', () => {
     const reducer = resourceReducer('hellos', {
       plugins: [
-        (state, action) => {
+        () => (state, action) => {
           if (action.type === 'SANDWICHES_ARE_GOOD') {
             return {
               ...state,
@@ -83,7 +95,7 @@ describe('reducer', function() {
 
           return state;
         },
-        (state, action) => {
+        () => (state, action) => {
           if (action.type === 'SANDWICHES_ARE_GOOD') {
             return {
               ...state,