+++ packages/resourceful-redux/test/unit/reducers/read.js
@@ -1,8 +1,220 @@
 import {resourceReducer, requestStatuses} from '../../../src';
 
 describe('reducers: read:', function() {
+  describe('READ_RESOURCES_PENDING:', () => {
+    it('should warn and not set a badly configured label', () => {
+      stub(console, 'error');
+      const reducer = resourceReducer('hellos', {
+        initialState: {
+          resources: {
+            1: {id: 1},
+            3: {id: 3},
+            4: {id: 4, lastName: 'camomile'},
+          },
+          labels: {},
+          meta: {
+            1: {
+              name: 'what'
+            },
+            3: {
+              deleteStatus: 'sandwiches'
+            },
+            4: {
+              createStatus: requestStatuses.SUCCEEDED,
+              selected: true
+            }
+          }
+        },
+        initialResourceMeta: {
+          selected: false,
+          createStatus: requestStatuses.PENDING
+        }
+      });
+
+      const reduced = reducer(undefined, {
+        type: 'READ_RESOURCES_PENDING',
+        resourceName: 'hellos',
+        label: true,
+        resources: [4, 5]
+      });
+
+      expect(reduced).to.deep.equal({
+        resources: {
+          1: {id: 1},
+          3: {id: 3},
+          4: {id: 4, lastName: 'camomile'}
+        },
+        labels: {},
+        meta: {
+          1: {
+            name: 'what'
+          },
+          3: {
+            deleteStatus: 'sandwiches'
+          },
+          4: {
+            selected: true,
+            createStatus: requestStatuses.SUCCEEDED,
+            readStatus: requestStatuses.PENDING,
+            updateStatus: requestStatuses.NULL,
+            deleteStatus: requestStatuses.NULL,
+          },
+          5: {
+            selected: false,
+            createStatus: requestStatuses.PENDING,
+            readStatus: requestStatuses.PENDING,
+            updateStatus: requestStatuses.NULL,
+            deleteStatus: requestStatuses.NULL,
+          }
+        }
+      });
+      expect(console.error.callCount).to.equal(1);
+    });
+
+    it('should return the correct state without erroring when called correctly', () => {
+      stub(console, 'error');
+      const reducer = resourceReducer('hellos', {
+        initialState: {
+          resources: {
+            1: {id: 1},
+            3: {id: 3},
+            4: {id: 4, lastName: 'camomile'},
+          },
+          labels: {},
+          meta: {
+            1: {
+              name: 'what'
+            },
+            3: {
+              deleteStatus: 'sandwiches'
+            },
+            4: {
+              createStatus: requestStatuses.SUCCEEDED,
+              selected: true
+            }
+          }
+        },
+        initialResourceMeta: {
+          selected: false,
+          createStatus: requestStatuses.PENDING
+        }
+      });
+
+      const reduced = reducer(undefined, {
+        type: 'READ_RESOURCES_PENDING',
+        resourceName: 'hellos',
+        resources: [4, 5]
+      });
+
+      expect(reduced).to.deep.equal({
+        resources: {
+          1: {id: 1},
+          3: {id: 3},
+          4: {id: 4, lastName: 'camomile'}
+        },
+        labels: {},
+        meta: {
+          1: {
+            name: 'what'
+          },
+          3: {
+            deleteStatus: 'sandwiches'
+          },
+          4: {
+            selected: true,
+            createStatus: requestStatuses.SUCCEEDED,
+            readStatus: requestStatuses.PENDING,
+            updateStatus: requestStatuses.NULL,
+            deleteStatus: requestStatuses.NULL,
+          },
+          5: {
+            selected: false,
+            createStatus: requestStatuses.PENDING,
+            readStatus: requestStatuses.PENDING,
+            updateStatus: requestStatuses.NULL,
+            deleteStatus: requestStatuses.NULL,
+          }
+        }
+      });
+      expect(console.error.callCount).to.equal(0);
+    });
+  });
+
   describe('READ_RESOURCES_SUCCEEDED:', () => {
-    it('returns the right state without a label, without IDs', () => {
+    it('warns when no resourceName is passed', () => {
+      stub(console, 'error');
+      const initialState = {
+        resources: {
+          1: {id: 1},
+          3: {id: 3},
+          4: {id: 4},
+        },
+        labels: {
+          pasta: {
+            hungry: true
+          }
+        },
+        meta: {
+          1: {
+            name: 'what'
+          },
+          3: {
+            deleteStatus: 'sandwiches'
+          }
+        }
+      };
+
+      const reducer = resourceReducer('hellos', {initialState});
+
+      const reduced = reducer(undefined, {
+        type: 'READ_RESOURCES_SUCCEEDED',
+        resources: [1, 2]
+      });
+
+      expect(reduced).to.deep.equal(initialState);
+      expect(console.error.callCount).to.equal(1);
+    });
+
+    it('warns when a resource _object_ is passed (not an array)', () => {
+      stub(console, 'error');
+      const initialState = {
+        resources: {
+          1: {id: 1},
+          3: {id: 3},
+          4: {id: 4},
+        },
+        labels: {
+          pasta: {
+            hungry: true
+          }
+        },
+        meta: {
+          1: {
+            name: 'what'
+          },
+          3: {
+            deleteStatus: 'sandwiches'
+          }
+        }
+      };
+
+      const reducer = resourceReducer('hellos', {initialState});
+
+      const reduced = reducer(undefined, {
+        type: 'READ_RESOURCES_SUCCEEDED',
+        resourceName: 'hellos',
+        resources: {
+          id: 20,
+          firstName: 'sandwiches'
+        }
+      });
+
+      expect(reduced).to.deep.equal(initialState);
+      expect(console.error.callCount).to.equal(1);
+    });
+
+    it('warns and returns the right state without a label, without IDs', () => {
+      stub(console, 'error');
       const initialState = {
         resources: {
           1: {id: 1},
@@ -32,9 +244,10 @@ describe('reducers: read:', function() {
       });
 
       expect(reduced).to.deep.equal(initialState);
+      expect(console.error.callCount).to.equal(1);
     });
 
-    it('returns state with resource object, no label, default options', () => {
+    it('returns state with resource array, no label, default options', () => {
       const reducer = resourceReducer('hellos', {
         initialState: {
           resources: {
@@ -51,6 +264,7 @@ describe('reducers: read:', function() {
               deleteStatus: 'sandwiches'
             },
             4: {
+              createStatus: requestStatuses.SUCCEEDED,
               selected: true
             }
           }
@@ -86,7 +300,7 @@ describe('reducers: read:', function() {
           },
           4: {
             selected: true,
-            createStatus: requestStatuses.NULL,
+            createStatus: requestStatuses.SUCCEEDED,
             readStatus: requestStatuses.SUCCEEDED,
             updateStatus: requestStatuses.NULL,
             deleteStatus: requestStatuses.NULL,
@@ -102,7 +316,7 @@ describe('reducers: read:', function() {
       });
     });
 
-    it('returns state with resource object, no label, mergeResources: false', () => {
+    it('returns state with resource array, no label, mergeResources: false', () => {
       const reducer = resourceReducer('hellos', {
         initialState: {
           resources: {
@@ -231,7 +445,93 @@ describe('reducers: read:', function() {
       });
     });
 
+    it('warns when a badly formatted label is passed in', () => {
+      stub(console, 'error');
+      const reducer = resourceReducer('hellos', {
+        initialState: {
+          resources: {
+            1: {id: 1},
+            3: {id: 3},
+            4: {id: 4, lastName: 'camomile'},
+          },
+          labels: {
+            sandwiches: {
+              ids: [1, 3],
+              status: requestStatuses.FAILED
+            },
+            pasta: {
+              ids: [4],
+              status: requestStatuses.PENDING
+            }
+          },
+          meta: {
+            1: {
+              name: 'what'
+            },
+            3: {
+              deleteStatus: 'sandwiches'
+            },
+            4: {
+              selected: true
+            }
+          }
+        }
+      });
+
+      const reduced = reducer(undefined, {
+        type: 'READ_RESOURCES_SUCCEEDED',
+        resourceName: 'hellos',
+        label: {},
+        resources: [
+          {id: 4, name: 'sandwiches'},
+          5
+        ]
+      });
+
+      expect(reduced).to.deep.equal({
+        resources: {
+          1: {id: 1},
+          3: {id: 3},
+          4: {id: 4, name: 'sandwiches', lastName: 'camomile'},
+          5: {id: 5}
+        },
+        labels: {
+          sandwiches: {
+            ids: [1, 3],
+            status: requestStatuses.FAILED
+          },
+          pasta: {
+            ids: [4],
+            status: requestStatuses.PENDING
+          }
+        },
+        meta: {
+          1: {
+            name: 'what'
+          },
+          3: {
+            deleteStatus: 'sandwiches'
+          },
+          4: {
+            selected: true,
+            createStatus: requestStatuses.NULL,
+            readStatus: requestStatuses.SUCCEEDED,
+            updateStatus: requestStatuses.NULL,
+            deleteStatus: requestStatuses.NULL,
+          },
+          5: {
+            createStatus: requestStatuses.NULL,
+            readStatus: requestStatuses.SUCCEEDED,
+            updateStatus: requestStatuses.NULL,
+            deleteStatus: requestStatuses.NULL,
+          }
+        }
+      });
+      expect(console.error.callCount).to.equal(1);
+    });
+
     it('returns state with resource object and label, ensuring no label ID dupes', () => {
+      stub(console, 'error');
       const reducer = resourceReducer('hellos', {
         initialState: {
           resources: {
@@ -312,6 +612,7 @@ describe('reducers: read:', function() {
           }
         }
       });
+      expect(console.error.callCount).to.equal(0);
     });
 
     it('returns state with resource object and label, ensuring empty label IDs works', () => {
@@ -551,6 +852,7 @@ describe('reducers: read:', function() {
     });
 
     it('returns state without a resource array, with a label', () => {
+      stub(console, 'error');
       const reducer = resourceReducer('hellos', {
         initialState: {
           resources: {
@@ -616,6 +918,8 @@ describe('reducers: read:', function() {
           }
         }
       });
+
+      expect(console.error.callCount).to.equal(1);
     });
   });
 });