+++ packages/resourceful-redux/test/unit/utils/reducer-generator.js
@@ -3,6 +3,8 @@ import {requestStatuses} from '../../../src';
 
 describe('reducerGenerator:', function() {
   it('passing no IDs and no label', () => {
+    stub(console, 'error');
+
     const state = {
       meta: {
         1: {
@@ -28,6 +30,7 @@ describe('reducerGenerator:', function() {
     const result = reducer(state, {});
 
     expect(result).to.equal(state);
+    expect(console.error.callCount).to.equal(1);
   });
 
   it('passing mixed resources array and no label', () => {
@@ -56,12 +59,22 @@ describe('reducerGenerator:', function() {
     const reducer = reducerGenerator('pasta', requestStatuses.FAILED);
     const result = reducer(state, {
       resources: [{id: 1}, {id: 5}, 6]
+    }, {
+      initialResourceMeta: {
+        selected: false,
+        readStatus: requestStatuses.PENDING
+      }
     });
 
     expect(result).to.deep.equal({
       meta: {
         1: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.PENDING,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED,
+          selected: false,
           hangry: true
         },
         2: {
@@ -69,9 +82,19 @@ describe('reducerGenerator:', function() {
           sandwichStatus: requestStatuses.PENDING
         },
         5: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.PENDING,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
+          selected: false,
           pastaStatus: requestStatuses.FAILED
         },
         6: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.PENDING,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
+          selected: false,
           pastaStatus: requestStatuses.FAILED
         }
       },
@@ -118,6 +141,10 @@ describe('reducerGenerator:', function() {
     expect(result).to.deep.equal({
       meta: {
         1: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED,
           hangry: true
         },
@@ -126,9 +153,17 @@ describe('reducerGenerator:', function() {
           sandwichStatus: requestStatuses.PENDING
         },
         5: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED
         },
         6: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED
         }
       },
@@ -227,6 +262,10 @@ describe('reducerGenerator:', function() {
     expect(result).to.deep.equal({
       meta: {
         1: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED,
           hangry: true
         },
@@ -235,9 +274,17 @@ describe('reducerGenerator:', function() {
           sandwichStatus: requestStatuses.PENDING
         },
         5: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED,
         },
         6: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED
         }
       },
@@ -286,6 +333,10 @@ describe('reducerGenerator:', function() {
     expect(result).to.deep.equal({
       meta: {
         1: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED,
         },
         2: {
@@ -293,9 +344,17 @@ describe('reducerGenerator:', function() {
           sandwichStatus: requestStatuses.PENDING
         },
         5: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED,
         },
         6: {
+          createStatus: requestStatuses.NULL,
+          readStatus: requestStatuses.NULL,
+          updateStatus: requestStatuses.NULL,
+          deleteStatus: requestStatuses.NULL,
           pastaStatus: requestStatuses.FAILED
         }
       },